+++ service/src/main/kotlin/io/provenance/explorer/domain/extensions/CoinExtensions.kt
@@ -1,12 +1,15 @@
 package io.provenance.explorer.domain.extensions
 
 import cosmos.base.v1beta1.CoinOuterClass
+import io.provenance.explorer.domain.models.explorer.CoinStr
 import java.math.BigDecimal
 
 
 const val NHASH = "nhash"
 
 fun BigDecimal.stringfy() = this.stripTrailingZeros().toPlainString()
+
+fun BigDecimal.toCoinStr(denom: String) = CoinStr(this.stringfy(), denom)
 
 fun String.toDecCoin() = BigDecimal(this.toBigInteger(), 18).stripTrailingZeros().toPlainString()